Updating file structure and enhace code

+++ core/helpers.py
@@ -1,9 +1,8 @@
 import os
 import json
 
-data_file = os.path.abspath("core/data_file.json")
-
-def if_datafile_not_exists():
+def ensure_datafile_exists():
+    data_file = os.path.abspath("data/data_file.json")
     if not os.path.exists(data_file):
         os.makedirs(os.path.dirname(data_file), exist_ok=True)
             
@@ -15,6 +14,34 @@ def if_datafile_not_exists():
         }
             
         with open(data_file, 'w') as f:
+            json.dump(default_data, f, indent=4)
+            
+        return default_data
+    
+def ensure_fileloc_exists():
+    data_file_loc = os.path.abspath("data/file_loc.json")
+    if not os.path.exists(data_file_loc):
+        os.makedirs(os.path.dirname(data_file_loc), exist_ok=True)
+            
+        default_data = {
+            "config_file_location": "data/data_file.json",
+            "prompt_file_location": "data/prompt.txt",
+            "memory_file_location": "data/memory.json"
+        }
+            
+        with open(data_file_loc, 'w') as f:
+            json.dump(default_data, f, indent=4)
+            
+        return default_data
+    
+def ensure_memoryfile_exists():
+    memory_file = os.path.abspath("data/memory.json")
+    if not os.path.exists(memory_file):
+        os.makedirs(os.path.dirname(memory_file), exist_ok=True)
+            
+        default_data = []
+            
+        with open(memory_file, 'w') as f:
             json.dump(default_data, f, indent=4)
             
         return default_data
